+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 0.11alpha
 - support for column grants
 - function ddlx_get_dependants_recursive() rolled into ddlx_get_dependants()
 - support for fdw options on columns
+- removed column 'kind' from ddlx_describe()
 
 Version 0.10
 ------------